+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 2.8.12)
+cmake_minimum_required(VERSION 3.5)
 project(bevfusion)
 
 add_definitions(-w)
@@ -24,62 +24,61 @@ set(CUDA_NVCC_FLAGS ${CUDA_NVCC_FLAGS}
 )
 
 # cuda
-set(CUDA_TOOLKIT_ROOT_DIR /usr/local/cuda-11.3)
+set(CUDA_TOOLKIT_ROOT_DIR /usr/local/cuda)
 set(CUDA_INSTALL_TARGET_DIR targets/x86_64-linux)
+
 set(CUDA_INCLUDE_DIRS ${CUDA_TOOLKIT_ROOT_DIR}/${CUDA_INSTALL_TARGET_DIR}/include)
 set(CUDA_LIBS ${CUDA_TOOLKIT_ROOT_DIR}/${CUDA_INSTALL_TARGET_DIR}/lib)
 
 # TENSORRT
 set(TensorRT_ROOT /home/lin/software/TensorRT-8.5.3.1)
-# set(TensorRT_ROOT ~/share/TensorRT-8.5.3.1)
+# set(TensorRT_ROOT ~/share/software/TensorRT-8.5.3.1)
 set(TensorRT_INCLUDE_DIRS ${TensorRT_ROOT}/include)
 set(TensorRT_LIBS ${TensorRT_ROOT}/lib/)
 
-
 find_package(CUDA REQUIRED)
 find_package(Protobuf REQUIRED)
-
 find_package(OpenCV REQUIRED)
-include_directories(${OpenCV_INCLUDE_DIRS})
-
 find_package(PCL REQUIRED)
-include_directories(${PCL_INCLUDE_DIRS})
 
-find_package(catkin REQUIRED COMPONENTS
-  roscpp
-  rospy
-  roslib
-  message_filters
+# ros依赖 find dependencies
+find_package(ament_cmake REQUIRED)
+find_package(rclpy REQUIRED)
+find_package(rclcpp REQUIRED)
+find_package(std_msgs REQUIRED)
+find_package(sensor_msgs REQUIRED)
+find_package(cv_bridge REQUIRED)
+find_package(message_filters REQUIRED)
 
-  sensor_msgs
-  pcl_conversions
-  pcl_ros 
-  cv_bridge
-)
+if(BUILD_TESTING)
+	find_package(ament_lint_auto REQUIRED)
+	ament_lint_auto_find_test_dependencies()
+endif()
+ament_package()
 
 include_directories(
+  ${TensorRT_INCLUDE_DIRS}
+  ${Protobuf_INCLUDE_DIRS}
+  ${CUDA_INCLUDE_DIRS}
+  ${OpenCV_INCLUDE_DIRS}
+  ${PCL_INCLUDE_DIRS}
+
   third_party/stb
-  third_party/pybind11/include
   third_party/cuOSD/src
   third_party/3DSparseConvolution/libspconv/include
   src/common
   src/
-  ${TensorRT_INCLUDE_DIRS}
-  ${Protobuf_INCLUDE_DIRS}
-  ${CUDA_INCLUDE_DIRS}
-  ${catkin_INCLUDE_DIRS}
+  
+  include_plugin
   include_ros
 )
 
-catkin_package()
+
 
 link_directories(
-  third_party/3DSparseConvolution/libspconv/lib/${arch}
-  build
-  /usr/lib/aarch64-linux-gnu
-  /usr/aarch64-linux-gnu/lib
   ${TensorRT_LIBS}
   ${CUDA_LIBS}
+  third_party/3DSparseConvolution/libspconv/lib/${arch}
 )
 
 file(GLOB_RECURSE CORE_FILES 
@@ -95,54 +94,44 @@ file(GLOB_RECURSE VISUALIZE_FILES
   src/common/visualize.cu
 )
 
-####################### libbevfusion_core.so ##########################
-cuda_add_library(bevfusion_core SHARED 
+
+####################### bevfusion program #############################
+
+cuda_add_executable(bevfusion_node
+  src_ros/main.cpp 
+  src_ros/bevfusion_node.cpp 
+  src_plugin/bevfusion_plugin.cpp
+  
+  ${VISUALIZE_FILES}
   ${CORE_FILES}
 )
 
-target_link_libraries(bevfusion_core
+target_link_libraries(bevfusion_node
+  libnvinfer.so
+  libspconv.so
+  ${Protobuf_LIBRARIES}
+  libnvinfer_plugin.so
+
   libcudart.so
   libcublasLt.so
-  libnvinfer.so
-  libspconv.so
-  ${Protobuf_LIBRARIES}
-  libnvinfer_plugin.so
-)
-
-####################### bevfusion program #############################
-# cuda_add_executable(demo
-#   src_ros/demo.cpp src_ros/bevfusion_plugin.cpp
-#   ${VISUALIZE_FILES}
-# )
-
-# target_link_libraries(demo
-#   bevfusion_core
-#   libnvinfer.so
-#   libspconv.so
-#   ${Protobuf_LIBRARIES}
-#   libnvinfer_plugin.so
-  
-#   ${OpenCV_LIBS}
-#   ${PCL_LIBRARIES}
-#   ${catkin_LIBRARIES}
-# )
-
-# bevfusion_node
-cuda_add_executable(bevfusion_node
-  src_ros/bevfusion_ros_node.cpp 
-  src_ros/bevfusion_ros.cpp 
-  src_ros/bevfusion_plugin.cpp
-  ${VISUALIZE_FILES}
-)
-
-target_link_libraries(bevfusion_node
-  bevfusion_core
-  libnvinfer.so
-  libspconv.so
-  ${Protobuf_LIBRARIES}
-  libnvinfer_plugin.so
   
   ${OpenCV_LIBS}
   ${PCL_LIBRARIES}
-  ${catkin_LIBRARIES}
 )
+
+ament_target_dependencies(bevfusion_node
+	rclcpp
+	std_msgs
+	sensor_msgs
+	cv_bridge
+  message_filters
+)
+
+
+install(TARGETS
+  bevfusion_node
+  DESTINATION lib/${PROJECT_NAME}
+)
+
+install(DIRECTORY launch
+  DESTINATION share/${PROJECT_NAME})
