+++ README.md
@@ -7,9 +7,9 @@ This repository contains source code and models for BEVFusion online real-time i
 
 # 1 依赖安装
 
-+ **`ubuntu-20.04,noetic,cuda-11.3, cudnn-8.6.0, TensorRT-8.5`**
++ **`ubuntu-20.04,galactic,cuda-11.3, cudnn-8.6.0, TensorRT-8.5`**
 
-1. 默认已安装`noetic, cuda, cudnn`, 已下载`TensorRT`源码
+1. 默认已安装`galactic, cuda, cudnn`, 已下载`TensorRT`源码
 
 2. `ros`依赖
 
@@ -19,10 +19,13 @@ mkdir -p bevfusion_ws/src
 
 # 2. 进入bevfusion_ws/src目录，拉取源码
 cd bevfusion_ws/src
-git clone https://github.com/linClubs/BEVFusion-ROS-TensorRT.git
+git clone https://github.com/linClubs/BEVFusion-ROS-TensorRT.git 
 
-# 3. 进入bevfusion_ws工作空间一键安装功能包需要ros依赖
-cd .. 
+# 3. 切换galactic-devel分支
+git branch galactic-devel
+
+# 4. 进入bevfusion_ws工作空间一键安装功能包需要ros依赖
+cd ../ 
 rosdep install -r -y --from-paths src --ignore-src --rosdistro $ROS_DISTRO
 ~~~
 
@@ -30,24 +33,28 @@ rosdep install -r -y --from-paths src --ignore-src --rosdistro $ROS_DISTRO
 
 4. [模型导出参考](https://github.com/NVIDIA-AI-IOT/Lidar_AI_Solution/blob/master/CUDA-BEVFusion/qat/README.md)
 
-+ 修改`./tool/environment.sh`中`cuda tensorrt cudnn`的路径, 运行`./tool/build_trt_engine.sh`生成tensorrt推理模型
-
-~~~python
-./tool/build_trt_engine.sh
-~~~
-
-
-5. `ros`包准备
+5. `rosbag`准备
 
 + `bevfusion`官方提供了已训练好的`nuscenes`模型参数
-
-+ rosbag数据转换参考[`nuscenes2rosbag`功能包](https://github.com/linClubs/nuscenes2rosbag)
-
 + `nuscenes`传感器之间的参数已给出,无需标定 
 
 如果需接真实的传感器进行场景测试,需提前完成**训练**和**标定**工作
 
 [传感器标定参考](https://github.com/linClubs/Calibration-Is-All-You-Need)
+
+
+[`nuscenes2rosbag`](https://github.com/linClubs/nuscenes2rosbag)
+
++ 上面生成的是`ros1`的`bag`，直接使用`rosbags`转化成`ros2`的`bag`
+
+~~~python
+# 1. install rosbags
+pip install rosbags
+
+# 2. ros1 bag convert to ros2 bag, result will be "nuscenes-103/"
+rosbags-convert nuscenes-103.bag
+~~~
+
 
 
 # 2 编译运行
@@ -63,7 +70,7 @@ set(CUDA_INCLUDE_DIRS ${CUDA_TOOLKIT_ROOT_DIR}/${CUDA_INSTALL_TARGET_DIR}/includ
 set(CUDA_LIBS ${CUDA_TOOLKIT_ROOT_DIR}/${CUDA_INSTALL_TARGET_DIR}/lib)
 
 # TENSORRT
-set(TensorRT_ROOT /home/lin/software/TensorRT-8.5.3.1)  # TensorRT修改这一行
+set(TensorRT_ROOT ~/software/TensorRT-8.5.3.1)  # TensorRT修改这一行
 # set(TensorRT_ROOT ~/share/TensorRT-8.5.3.1)           
 set(TensorRT_INCLUDE_DIRS ${TensorRT_ROOT}/include)
 set(TensorRT_LIBS ${TensorRT_ROOT}/lib/)
@@ -72,29 +79,51 @@ set(TensorRT_LIBS ${TensorRT_ROOT}/lib/)
 
 2. 编译运行
 
-+ `bevfusion_node.launch`修改`model_name`与`precision`参数值
-
-`model_name: resnet50/resnet50int8/swint`
-`precision:  fp16/int8`
-`swint + int8`模式不能工作
++ `bevfusion.launch.py`修改`model_name`与`precision`参数值
+~~~python
+# model_name: resnet50/resnet50int8/swint
+# precision:  fp16/int8
+# swint + int8模式不能工作
+parameters=[
+			{'model_name': 'resnet50'},
+			{'precision' : 'int16'}
+		]
+~~~
 
 ~~~python
 # 1. 编译
-catkin_make
+colcon build --symlink-install
 
 # 2. source工作空间
-source devel/setup.bash
+source install/setup.bash
 
 # 3. 运行bevfusion_node
-roslaunch bevfusion bevfusion_node.launch
+roslaunch bevfusion bevfusion.launch.py
 
 # 4. 播放数据集
- rosbag play 103.bag 
+ rosbag play nuscenes-103.db3
 ~~~
 
-3. 运行报错`tool/simhei.ttf`找不到, 全局搜索`tool/simhei.ttf`或者`UseFont`关键字
+---
 
-在`/src/common/visualize.cu`中修改`UseFont`的值即可,改成`simhei.ttf`正确的路径即可
+
+3. 错误修改
++ 报错1 运行报错`tool/simhei.ttf`找不到, 全局搜索`tool/simhei.ttf`或者`UseFont`关键字
+
+修改：在`/src/common/visualize.cu`中修改`UseFont`的值即可,改成`simhei.ttf`正确的路径即可
+
+
++ 报错2 运行`ros2 run bevfusion bevfusion_node` 
+`error while loading shared libraries: libspconv.so: cannot open shared object file: No such file or directory`
+
+修改：
+~~~python
+# 查看LD_LIBRARY_PATH路径
+echo $LD_LIBRARY_PATH
+
+# 如果没有libspconv.so路径就按照下面代码添加即可
+export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/path/to/third_party/3DSparseConvolution/libspconv/lib/x86_64
+~~~
 
 ---
