+++ src/common/visualize.cu
@@ -35,7 +35,9 @@
 #include "visualize.hpp"
 
 namespace nv {
-#define UseFont "./tool/simhei.ttf"
+// #define UseFont "./tool/simhei.ttf"
+#define UseFont "/home/lin/ros2_code/bevfusion_ws/src/FastBEV-ROS-TensorRT/tool/simhei.ttf"
+
 #define MaxDistance 50
 #define INTER_RESIZE_COEF_BITS 11
 #define INTER_RESIZE_COEF_SCALE (1 << INTER_RESIZE_COEF_BITS)
